Rename generated provider files with resource name
